feat(navigation): Refactor all code with token bearer and clean code + navigation between screen + update user profile info screen  with the backend call
